Add basic test and simplify the rendering
